Add spec about html comments inside of markup.
HTML comments are more complex then just `<!-- something -->`:
http://spec.commonmark.org/0.22/#example-571

Working on my petproject allmychanges.com, I was investigating
my angular's changelog parsing was broken after the switch
from python-markdown2 to CommonMark.

An I found that a problem is this line of text:

```
s <!--> 1234567890123
[a link]()
```

Parsing of this small snippet takes about 15 seconds on my VPS. And each
symbol after the HTML comment multiplies this time by 2.

I've checked, that HTML comment shouldn't be started from the begging, that is why
there is  "s " before it.

And a link should be present in same paragraph, if you separate the link with empy
line, everything is ok.

By the way, full part of angular's changelog is:

```
   - sanitizer should not accept <!--> as a valid comment
    ([21e9e8cf](https://github.com/angular/angular.js/commit/21e9e8cf68ef007136da6cc212d2f1f252fb668a))
```

So, this is real example.