Only use one readme file
It doesn't make sense to have to maintain two readme files. If we
use the rst file, then it should get rendered correctly on
pypi here: https://pypi.python.org/pypi/CommonMark
Which is currently failing to render the markdown file.

We are using README.md as a test case, so I've moved it to
CommonMark/test/test.md.