Add normalize() method for AST nodes
The parser produces an AST, which may contain multiple
text nodes in a row, which is difficult to work with.

Example commonmark input:

_unfinished emphasis

For this input the parser produces 2 separate text nodes
for "_" and "unfinished emphasis".

The normalize() method on an AST node will walk through
the tree and concatenate these consecutive text nodes.