Support rendering multi-line list items to rst
For list items going across multiple lines, the renderer previously did
not indent the latter lines, which leads to invalid rst (see
https://docutils.sourceforge.io/docs/ref/rst/restructuredtext.html#bullet-lists;
subsequent list items must line up with the first item).

This commit fixes that by introducing an indent when entering a list
item.

Note that there are still some issues (for example, paragraphs within
list items are still not handled correctly), but this at least seems to
be a step in the right direction.